% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adding.R
\name{Adding}
\alias{Adding}
\title{Adding}
\usage{
Adding(x, y)
}
\arguments{
\item{x}{is the first number to be added}

\item{y}{is the second number to be added}
}
\value{
sum of two numbers
}
\description{
This function adds two numbers together
}
\examples{
Adding(x=4,y=7)
}
\author{
Melaina Wright
}
