% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{cor_matrix}
\alias{cor_matrix}
\title{A HW5Package Function}
\usage{
cor_matrix(year)
}
\arguments{
\item{year}{no default}
}
\description{
Outputs a correlation matrix for all numerical data in a given year
}
\examples{
cor_matrix()
}
\keyword{correlation}
