% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1Pythagorean.R
\name{f1Pythagorean}
\alias{f1Pythagorean}
\title{Function of Pythagorean theorem}
\usage{
f1Pythagorean(a = NULL, b = NULL, c = NULL)
}
\arguments{
\item{a}{the length of right-angle side}

\item{b}{the length of right-angle side}

\item{c}{the length of side opposite the right angle}
}
\value{
the length of third side of this triangle
}
\description{
This function is used calculating the length of the third side,
given the lengths of two sides of the triangle.
}
\examples{
f1Pythagorean(a=3,b=4)

}
\keyword{length}
\keyword{of}
\keyword{triangle}
