% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2trimmed_mean.R
\name{f2trimmed_mean}
\alias{f2trimmed_mean}
\title{Function of trimmed mean}
\usage{
f2trimmed_mean(num_vec, s_smallest, l_largest)
}
\arguments{
\item{num_vec}{the numeric vector}

\item{s_smallest}{number of smallest numbers ignored}

\item{l_largest}{number of largest numbers ignored}
}
\value{
the calculated trimmed mean
}
\description{
This function is used to calculate trimmed mean
}
\examples{
f2trimmed_mean(c(1,7,3,2,5,0.5,9,10),1,2)

}
\keyword{mean}
\keyword{trimmed}
