% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries_summary.R
\name{fisheries_summary}
\alias{fisheries_summary}
\title{fisheries_summary}
\usage{
fisheries_summary(fishprice, fishnumber, graph = "true")
}
\arguments{
\item{fishprice}{is a table that has prices for different fish}

\item{fishnumber}{is a table that has the number caught for each fish species for each location. Each location is in a different column. Each fish is in a different row.}
}
\value{
most frequently caught fish in each location, total revenue for each location, total fisheries sum, and graph revenue by location and add total to the plot if the user requests it.
}
\description{
This function computes the most frequently caught fish in each location, total revenue for each location, total fisheries sum, and graph revenue by location and add total to the plot if the user requests it.
}
\examples{
fisheries_summary(fishprice=fishprice, fishnumber=fishnumber, graph="true")
}
\author{
Melaina Wright
}
