% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnumber.R
\docType{data}
\name{fishnumber}
\alias{fishnumber}
\title{fishnumber}
\format{it is a .csv file with 4 rows and 4 columns with four variables:
\itemize{
\item species is the fish species
\item A is the number of fish caught of each species from location A
#' \item B is the number of fish caught of each species from location B
#' \item C is number of fish caught of each species from location C
}}
\usage{
fishnumber
}
\description{
This mock data was created by Melaina Wright to be used to demonstrate the fisheries_summary function.
}
\keyword{datasets}
