% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishprice.R
\docType{data}
\name{fishprice}
\alias{fishprice}
\title{fishprice}
\format{it is a .csv file with 4 rows and 2 columns with two variables:
\itemize{
\item species is the fish species
\item price is the market price of the fish per kg
}}
\usage{
fishprice
}
\description{
This mock data was created by Melaina Wright to be used to demonstrate the fisheries_summary function.
}
\keyword{datasets}
