% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{highest_ft}
\alias{highest_ft}
\title{A HW5Package Function}
\usage{
highest_ft(year)
}
\arguments{
\item{year}{no default}
}
\description{
Outputs all stats for the player with the highest free throw number in a given year
}
\examples{
highest_ft()
}
\keyword{free}
\keyword{throws}
