% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{oldest_player}
\alias{oldest_player}
\title{A HW5Package Function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{no default}
}
\description{
Outputs all stats for the oldest player in a given year
Returns first result in alphabetical order if there is a tie
}
\examples{
oldest_player()
}
\keyword{age}
