% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pb.R
\name{Pb}
\alias{Pb}
\title{Pb}
\usage{
Pb(m, V, A, cRolling = 0.015, cDrag = 0.3, pAir = 1.2, g = 9.8)
}
\arguments{
\item{m}{mass of car inkg}

\item{V}{velocity of car in m/s}

\item{A}{surface area of car in m2}

\item{cRolling}{rolling aerodynamic resistive coefficient, default is 0.015}

\item{cDrag}{drag aerodynamic resistive coefficient, default is 0.3}

\item{pAir}{density of air, default is 1.2kg/m3}

\item{g}{acceleration due to gravity is 9.8m/s2}
}
\value{
Power
}
\description{
This function computes automobile fuel efficiency is as the power required to keep a car moving at a given speed
}
\examples{
Pb(31752, 20, 60)
}
\author{
Melaina Wright
}
